+++ tests/core/arc.cpp
@@ -50,6 +50,6 @@ TEST(core, arc_plain) {
     /// Checksum test against reference
     std::string test_name =
         ::testing::UnitTest::GetInstance()->current_test_info()->name();
-    std::size_t file_checksum = of.checksum();
+    std::uint32_t file_checksum = of.checksum();
     EXPECT_TRUE(test::checksum(test_name, file_name, file_checksum));
 }
