+++ tests/core/bezier.cpp
@@ -50,7 +50,7 @@ TEST(core, bezier_single_segment) {
     // Checksum test against reference
     std::string test_name =
         ::testing::UnitTest::GetInstance()->current_test_info()->name();
-    std::size_t file_checksum = of.checksum();
+    std::uint32_t file_checksum = of.checksum();
     EXPECT_TRUE(test::checksum(test_name, file_name, file_checksum));
 }
 
@@ -96,6 +96,6 @@ TEST(core, bezier_spiral) {
     // Checksum test against reference
     std::string test_name =
         ::testing::UnitTest::GetInstance()->current_test_info()->name();
-    std::size_t file_checksum = of.checksum();
+    std::uint32_t file_checksum = of.checksum();
     EXPECT_TRUE(test::checksum(test_name, file_name, file_checksum));
 }
