+++ tests/core/gradient.cpp
@@ -60,7 +60,7 @@ TEST(core, gradient_box_linear_x) {
     // Checksum test against reference
     std::string test_name =
         ::testing::UnitTest::GetInstance()->current_test_info()->name();
-    std::size_t file_checksum = mfile.checksum();
+    std::uint32_t file_checksum = mfile.checksum();
     EXPECT_TRUE(test::checksum(test_name, file_name, file_checksum));
 }
 
@@ -94,7 +94,7 @@ TEST(core, gradient_box_vertical) {
     // Checksum test against reference
     std::string test_name =
         ::testing::UnitTest::GetInstance()->current_test_info()->name();
-    std::size_t file_checksum = mfile.checksum();
+    std::uint32_t file_checksum = mfile.checksum();
     EXPECT_TRUE(test::checksum(test_name, file_name, file_checksum));
 }
 
@@ -128,7 +128,7 @@ TEST(core, gradient_box_vertical_label) {
     // Checksum test against reference
     std::string test_name =
         ::testing::UnitTest::GetInstance()->current_test_info()->name();
-    std::size_t file_checksum = mfile.checksum();
+    std::uint32_t file_checksum = mfile.checksum();
     EXPECT_TRUE(test::checksum(test_name, file_name, file_checksum));
 }
 
@@ -162,7 +162,7 @@ TEST(core, gradient_box_horizontal) {
     // Checksum test against reference
     std::string test_name =
         ::testing::UnitTest::GetInstance()->current_test_info()->name();
-    std::size_t file_checksum = mfile.checksum();
+    std::uint32_t file_checksum = mfile.checksum();
     EXPECT_TRUE(test::checksum(test_name, file_name, file_checksum));
 }
 
@@ -197,6 +197,6 @@ TEST(core, gradient_box_horizontal_label) {
     // Checksum test against reference
     std::string test_name =
         ::testing::UnitTest::GetInstance()->current_test_info()->name();
-    std::size_t file_checksum = mfile.checksum();
+    std::uint32_t file_checksum = mfile.checksum();
     EXPECT_TRUE(test::checksum(test_name, file_name, file_checksum));
 }
